requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
google-api-python-client==2.122.0
groq==0.4.0
ccxt==4.1.68
httpx==0.23.0
